Small fixes to match Siggy and the class diagram
Sorry Siggy, I was wrong about creating a folder in save() method.
Python can't directly create a new folder for saving data if the folder
doesn't exist. The folder must exist. But as long as it exists, it can
deal with any folders.
And I have been looking for the way to to run the py file by simple
typing "finename.py" in commandline, but I can't find a solution. I
guess it's ok since it's kinda beyond of marking schedule...